Icon generation before build time - Moved icon code into gen-icons.ts   - It generates icons.ts   - gets called by `npm run gen-icons`
