+++ src/ide/icons.ts
@@ -1,516 +1,52 @@
 "use strict";
-
-import { SVGDrawingContext } from "./svg";
-
-export const icons = (function () {
-	let icons: any = {};
-
-	// Gui icons
-
-	icons.window = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-panel-back");
-		draw.rect(2.5, 2.5, 15, 15);
-
-		draw.setClass("icon-panel-title");
-		draw.rect(2.5, 2.5, 15, 3);
-
-		draw.setClass("icon-border");
-		draw.rect(2.5, 2.5, 15, 15);
-	};
-
-	icons.dockLeft = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke: #666; fill: none");
-		draw.rect(2.5, 2.5, 13, 13);
-
-		draw.setClass("icon-panel-back");
-		draw.rect(2.5, 2.5, 7, 13);
-
-		draw.setClass("icon-panel-title");
-		draw.rect(2.5, 2.5, 7, 3);
-
-		draw.setClass("icon-border");
-		draw.rect(2.5, 2.5, 7, 13);
-	};
-
-	icons.dockRight = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke: #666; fill: none");
-		draw.rect(2.5, 2.5, 13, 13);
-
-		draw.setClass("icon-panel-back");
-		draw.rect(8.5, 2.5, 7, 13);
-
-		draw.setClass("icon-panel-title");
-		draw.rect(8.5, 2.5, 7, 3);
-
-		draw.setClass("icon-border");
-		draw.rect(8.5, 2.5, 7, 13);
-	};
-
-	icons.maximize = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-panel-back");
-		draw.rect(2.5, 2.5, 13, 13);
-
-		draw.setClass("icon-panel-title");
-		draw.rect(2.5, 2.5, 13, 3);
-
-		draw.setClass("icon-border");
-		draw.rect(2.5, 2.5, 13, 13);
-	};
-
-	icons.float = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke: #666; fill: none");
-		draw.rect(2.5, 2.5, 13, 13);
-
-		draw.setClass("icon-panel-back");
-		draw.rect(4.5, 5.5, 9, 8);
-
-		draw.setClass("icon-panel-title");
-		draw.rect(4.5, 5.5, 9, 3);
-
-		draw.setClass("icon-border");
-		draw.rect(4.5, 5.5, 9, 8);
-	};
-
-	icons.minimize = function (draw: SVGDrawingContext) {
-		draw.setClassAndStyle("icon-border", "stroke-width: 2.5");
-		draw.line(3, 14.5, 15, 14.5);
-	};
-
-	icons.help = function (draw: SVGDrawingContext) {
-		draw.setClassAndStyle("icon-border", "stroke-width: 2");
-
-		draw.beginPath();
-		draw.arc(9, 6, 4, 1 * Math.PI, 2.25 * Math.PI);
-		draw.lineTo(9, 11.5);
-		draw.lineTo(9, 13);
-		draw.endPath();
-
-		draw.line(9, 15, 9, 17);
-	};
-
-	icons.close = function (draw: SVGDrawingContext) {
-		draw.setClassAndStyle("icon-border", "stroke-width: 2");
-		draw.line(4, 4, 14, 14);
-		draw.line(4, 14, 14, 4);
-	};
-
-	// Message box icons, larger in size
-
-	icons.msgBoxQuestion = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke-width: 2; stroke: #04d; fill: #16f");
-		draw.circle(20, 20, 18.5);
-
-		draw.setStyle("stroke-width: 2; stroke: #fff; fill: none");
-		draw.beginPath();
-		draw.arc(20, 15, 7, 1 * Math.PI, 2.25 * Math.PI);
-		draw.lineTo(20, 25);
-		draw.lineTo(20, 28);
-		draw.endPath();
-
-		draw.line(20, 31, 20, 34);
-	};
-
-	icons.msgBoxExclamation = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke-width: 2; stroke: #a91; fill: #ec2");
-		draw.polygon([19, 2], [21, 2], [38, 36], [37, 38], [3, 38], [2, 36]);
-
-		draw.setStyle("stroke-width: 2; stroke: #000; fill: none");
-		draw.line(20, 10, 20, 28);
-		draw.line(20, 31, 20, 34);
-	};
-
-	// Ide specific icons
-
-	// Toolbar icons
-	// icon parts used several times are written as a function,
-
-	function draw_icon_paper(draw: SVGDrawingContext) {
-		draw.setStyle("stroke: #333; fill: #fff");
-		draw.polygon(
-			[4.5, 7.5],
-			[8.5, 3.5],
-			[14.5, 3.5],
-			[14.5, 16.5],
-			[4.5, 16.5]
-		);
-
-		draw.setStyle("stroke: #333; fill: none");
-		draw.polyline([4.5, 7.5], [8.5, 7.5], [8.5, 3.5]);
-	}
-
-	function draw_icon_floppy_disk(draw: SVGDrawingContext) {
-		draw.setStyle("stroke: #139; fill: #36d");
-		draw.polygon(
-			[3.5, 3.5],
-			[16.5, 3.5],
-			[16.5, 16.5],
-			[5.5, 16.5],
-			[3.5, 14.5]
-		);
-
-		draw.setStyle("stroke: none; fill: #eef");
-		draw.rect(7, 11, 6, 5);
-
-		draw.setStyle("stroke: none; fill: #36d");
-		draw.rect(8, 12, 2, 3);
-
-		draw.setStyle("stroke: none; fill: #fff");
-		draw.rect(6, 4, 8, 5);
-	}
-
-	function draw_icon_pencil_overlay(draw: SVGDrawingContext) {
-		// draw pencil
-		// shadow
-		draw.setStyle("stroke-width: 3; stroke: #0005; fill: none");
-		draw.line(8, 8, 8 + 10, 8 + 10);
-
-		draw.setStyle("stroke: none; fill: #fc9");
-		draw.polygon([8, 6], [11, 7], [9, 9]);
-
-		draw.setStyle("stroke: none; fill: #000");
-		draw.polygon([8, 6], [9.5, 6.5], [8.5, 7.5]);
-
-		draw.setStyle("stroke-width: 3; stroke: #000; fill: none");
-		draw.line(10, 8, 18, 16);
-
-		draw.setStyle("stroke-width: 2; stroke: #dd0; fill: none");
-		draw.line(10, 8, 18, 16);
-
-		draw.setStyle("stroke: none; fill: #000");
-		draw.polygon([8, 6], [9.5, 6.5], [8.5, 7.5]);
-		draw.beginPath();
-		draw.arc(18, 16, 1.5, 1.75 * Math.PI, 2.75 * Math.PI);
-		draw.endPath();
-
-		draw.setStyle("stroke: none; fill: #f33");
-		draw.circle(18, 16, 1);
-	}
-
-	icons.newDocument = function (draw: SVGDrawingContext) {
-		draw_icon_paper(draw);
-
-		draw.setStyle("stroke: #030; fill: #0a0");
-		draw.circle(14, 14, 4.75);
-
-		draw.setStyle("stroke-width: 2; stroke: #fff; fill: none");
-		draw.line(14, 17, 14, 11);
-		draw.line(11, 14, 17, 14);
-	};
-
-	icons.openDocument = function (draw: SVGDrawingContext) {
-		draw.setStyle("fill: #ec5; stroke: #330");
-		draw.polygon(
-			[2.5, 4.5],
-			[7.5, 4.5],
-			[9.5, 6.5],
-			[15.5, 6.5],
-			[15.5, 15.5],
-			[2.5, 15.5]
-		);
-
-		draw.setStyle("fill: #fd6; stroke: #330");
-
-		draw.polygon([5.5, 8.5], [17.5, 8.5], [15.5, 15.5], [3.5, 15.5]);
-	};
-
-	icons.saveDocument = function (draw: SVGDrawingContext) {
-		draw_icon_floppy_disk(draw);
-	};
-
-	icons.saveDocumentAs = function (draw: SVGDrawingContext) {
-		draw_icon_floppy_disk(draw);
-		draw_icon_pencil_overlay(draw);
-	};
-
-	icons.run = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-green-fill");
-		draw.polygon([5, 5], [15, 10], [5, 15]);
-	};
-
-	icons.interrupt = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-red-fill");
-		draw.rect(5, 5, 4, 10);
-		draw.rect(11, 5, 4, 10);
-	};
-
-	icons.reset = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-red-fill");
-		draw.rect(5, 5, 10, 10);
-	};
-
-	icons.stepInto = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-neutral-fill");
-		draw.rect(10, 3, 7, 2);
-		draw.rect(13, 6, 4, 2);
-		draw.rect(13, 9, 4, 2);
-		draw.rect(13, 12, 4, 2);
-		draw.rect(10, 15, 7, 2);
-
-		draw.setClass("icon-blue-line");
-		draw.polyline([8, 4], [3, 4], [3, 10], [6, 10]);
-
-		draw.setClass("icon-blue-fill");
-		draw.polygon([5, 7], [5, 13], [9.5, 10]);
-	};
-
-	icons.stepOver = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-neutral-fill");
-		draw.rect(10, 3, 7, 2);
-		draw.rect(13, 6, 4, 2);
-		draw.rect(13, 9, 4, 2);
-		draw.rect(13, 12, 4, 2);
-		draw.rect(10, 15, 7, 2);
-
-		draw.setClass("icon-blue-line");
-		draw.polyline([8, 4], [3, 4], [3, 16], [6, 16]);
-
-		draw.setClass("icon-blue-fill");
-		draw.polygon([5, 13], [5, 19], [9.5, 16]);
-	};
-
-	icons.stepOut = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-neutral-fill");
-		draw.rect(10, 3, 7, 2);
-		draw.rect(13, 6, 4, 2);
-		draw.rect(13, 9, 4, 2);
-		draw.rect(13, 12, 4, 2);
-		draw.rect(10, 15, 7, 2);
-
-		draw.setClass("icon-blue-line");
-		draw.polyline([11, 10], [3, 10], [3, 16], [6, 16]);
-
-		draw.setClass("icon-blue-fill");
-		draw.polygon([5, 13], [5, 19], [9.5, 16]);
-	};
-
-	icons.breakPoint = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-red-fill");
-		draw.circle(10, 10, 3.9);
-	};
-
-	/*icons.search = function(draw: SVGDrawingContext) {
-		draw.setClassAndStyle("icon-neutral-line", "stroke-width: 1.5");
-		draw.beginPath();
-		draw.arc(8, 8, 5, 0.25*Math.PI, 2.25*Math.PI, false); // starting/ending point at 45 degrees south-east
-		draw.lineTo(17, 17);
-		draw.endPath();
-	};*/
-
-	icons.export = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-green-frame");
-		draw.polygon(
-			[3, 7],
-			[10, 7],
-			[10, 3],
-			[17, 10],
-			[10, 17],
-			[10, 13],
-			[3, 13]
-		);
-	};
-
-	icons.config = function (draw: SVGDrawingContext) {
-		draw.setClass("icon-neutral-fill");
-		draw.circle(10, 10, 2.0);
-
-		draw.setClassAndStyle("icon-neutral-line", "stroke-width: 2");
-		draw.circle(10, 10, 5.7);
-
-		for (let i = 0; i < 12; i++) {
-			let a = ((i + 0.5) * Math.PI) / 6;
-			draw.line(
-				10 + 6.0 * Math.cos(a),
-				10 + 6.0 * Math.sin(a),
-				10 + 9.4 * Math.cos(a),
-				10 + 9.4 * Math.sin(a)
-			);
-		}
-	};
-
-	icons.restorePanels = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke: #aaa; fill: #fff");
-		draw.rect(2.5, 2.5, 15, 15);
-
-		draw.setStyle("stroke: none; fill: #ccc");
-		draw.rect(11, 3, 1, 14);
-
-		draw.setStyle("stroke: none; fill: #77f");
-		draw.rect(3, 3, 14, 1);
-		draw.rect(3, 12, 8, 1);
-
-		draw.setStyle("stroke: #222; stroke-width: 1.7; fill: none");
-		draw.beginPath();
-		draw.arc(9.5, 10.5, 4, 1.25 * Math.PI, 2.6 * Math.PI);
-		draw.endPath();
-
-		draw.setStyle("stroke: none; fill: #222");
-		draw.polygon([5, 5], [5, 10], [10, 10]);
-	};
-
-	// Panel icons
-	icons.editor = function (draw: SVGDrawingContext) {
-		draw_icon_paper(draw);
-
-		draw.setStyle("stroke: none; fill: #777");
-		draw.rect(10, 5, 3, 1);
-		draw.rect(10, 7, 2, 1);
-		draw.rect(7, 9, 4, 1);
-		draw.rect(6, 11, 7, 1);
-		draw.rect(9, 13, 4, 1);
-	};
-
-	icons.messages = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke: #222; fill: #fff");
-		draw.beginPath();
-		draw.ellipse(9.5, 8.5, 7, 6, 0, 0.73 * Math.PI, 2.57 * Math.PI);
-		draw.lineTo(4.5, 17);
-		draw.closePath();
-		draw.endPath();
-
-		draw.setStyle("stroke: none; fill: #777");
-		draw.rect(8, 6, 3, 1);
-		draw.rect(6, 8, 2, 1);
-		draw.rect(9, 8, 5, 1);
-		draw.rect(7, 10, 4, 1);
-	};
-
-	icons.stackView = function (draw: SVGDrawingContext) {
-		// white top
-		draw.setStyle("stroke: none; fill: #fff");
-		draw.polygon([4, 5.5], [10, 8.5], [16, 5.5], [10, 2.5]);
-
-		// shaded lower pages
-		draw.setStyle("stroke: none; fill: #bbb");
-		draw.polygon([4, 5.5], [4, 14.5], [10, 17.5], [10, 8.5]);
-
-		draw.setStyle("stroke: none; fill: #999");
-		draw.polygon([10, 17.5], [16, 14.5], [16, 5.5], [10, 8.5]);
-
-		draw.setStyle("stroke-width: 0.6; stroke: #222; fill: none");
-		for (let i = 8; i < 17; i += 3) {
-			draw.polyline([3, i + 0.5], [10, i + 3.5], [17, i + 0.5]);
-		}
-
-		// top frame
-		draw.polygon(
-			[3.5, 5.3],
-			[3.5, 5.7],
-			[10, 8.5],
-			[16.5, 5.7],
-			[16.5, 5.3],
-			[10, 2.5]
-		);
-	};
-
-	icons.programView = function (draw: SVGDrawingContext) {
-		// Outline
-		draw.setStyle("stroke:none; fill: #eeeeeec0");
-		draw.polygon(
-			[3, 2],
-			[13, 2],
-			[13, 5],
-			[15, 5],
-			[15, 8],
-			[17, 8],
-			[17, 12],
-			[13, 12],
-			[13, 14],
-			[15, 14],
-			[15, 18],
-			[5, 18],
-			[5, 15],
-			[3, 15],
-			[3, 11],
-			[7, 11],
-			[7, 9],
-			[5, 9],
-			[5, 6],
-			[3, 6]
-		);
-
-		// Black boxes
-		draw.setStyle("stroke:none; fill: #000");
-		draw.rect(4, 3, 8, 2);
-		draw.rect(6, 6, 8, 2);
-		draw.rect(8, 9, 8, 2);
-		draw.rect(4, 12, 8, 2);
-		draw.rect(6, 15, 8, 2);
-	};
-
-	icons.turtle = function (draw: SVGDrawingContext) {
-		// draws literally a turtle
-
-		draw.setStyle("stroke: #070; fill: #2c3");
-
-		// head
-		draw.beginPath();
-		draw.ellipse(9.5, 5, 2, 2.5, 0, 0 * Math.PI, 2 * Math.PI);
-		draw.closePath();
-		draw.endPath();
-
-		// legs
-		draw.setStyle("stroke-width: 1.6; stroke: #070; fill: none");
-
-		draw.polyline([3.5, 6], [9.5, 11], [15.5, 6]);
-
-		draw.polyline([4.5, 17], [9.5, 10], [14.5, 17]);
-
-		// tail
-		draw.setStyle("stroke-width: 1.3; stroke: #070; fill: none");
-		draw.line(9.5, 17, 8.5, 19);
-
-		// main body
-		draw.setStyle("stroke: #070; fill: #2c3");
-		draw.beginPath();
-		draw.ellipse(9.5, 11.5, 4, 5, 0, 0 * Math.PI, 2 * Math.PI);
-		draw.closePath();
-		draw.endPath();
-
-		draw.setStyle("stroke: #0709; fill: none");
-		draw.beginPath();
-		draw.ellipse(8.8, 10.8, 4, 5, 0, -0.3 * Math.PI, 0.7 * Math.PI);
-		draw.endPath();
-	};
-
-	icons.canvas = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke: none; fill: #333");
-		draw.rect(3, 2, 14, 16);
-
-		draw.setStyle("stroke: none; fill: #fff");
-		draw.rect(4, 3, 12, 14);
-
-		draw.setStyle("stroke: #00cc; fill: #00c8");
-		draw.rect(5.5, 5.5, 6, 6);
-
-		draw.setStyle("stroke: #c00c; fill: #c008");
-		draw.circle(11, 11, 3.5);
-	};
-
-	/*icons.tutorial = function (draw: SVGDrawingContext) {
-		draw.setStyle("stroke-width: 3; stroke: #666; fill: none");
-
-		draw.beginPath();
-		draw.moveTo(5.0, 20);
-		draw.lineTo(9.0, 2);
-		draw.lineTo(11.0, 2);
-		draw.lineTo(15.0, 20);
-		draw.endPath();
-
-		draw.setStyle("stroke-width: 2; stroke: #db7; fill: none");
-		draw.polyline([5, 20], [9, 2], [11, 2], [15, 20]);
-
-		draw.setStyle("stroke: none; fill: #666");
-		draw.rect(1.5, 2.5, 17, 12);
-
-		draw.setStyle("stroke: none; fill: #db7");
-		draw.rect(2, 3, 16, 11);
-
-		draw.setStyle("stroke: none; fill: #000");
-		draw.rect(3, 4, 14, 9);
-
-		draw.setStyle("stroke: #fff; fill: none");
-		draw.polyline([5, 6], [7, 8], [5, 10]);
-	};*/
-
-	return icons;
-})();
+// Auto generated by 'gen-icons.ts'
+export const icons = {
+	window: '<rect x="2.5" y="2.5" width="15" height="15" class="icon-panel-back"/><rect x="2.5" y="2.5" width="15" height="3" class="icon-panel-title"/><rect x="2.5" y="2.5" width="15" height="15" class="icon-border"/>',
+	dockLeft:
+		'<rect x="2.5" y="2.5" width="13" height="13" style="stroke: #666; fill: none"/><rect x="2.5" y="2.5" width="7" height="13" class="icon-panel-back"/><rect x="2.5" y="2.5" width="7" height="3" class="icon-panel-title"/><rect x="2.5" y="2.5" width="7" height="13" class="icon-border"/>',
+	dockRight:
+		'<rect x="2.5" y="2.5" width="13" height="13" style="stroke: #666; fill: none"/><rect x="8.5" y="2.5" width="7" height="13" class="icon-panel-back"/><rect x="8.5" y="2.5" width="7" height="3" class="icon-panel-title"/><rect x="8.5" y="2.5" width="7" height="13" class="icon-border"/>',
+	maximize:
+		'<rect x="2.5" y="2.5" width="13" height="13" class="icon-panel-back"/><rect x="2.5" y="2.5" width="13" height="3" class="icon-panel-title"/><rect x="2.5" y="2.5" width="13" height="13" class="icon-border"/>',
+	float: '<rect x="2.5" y="2.5" width="13" height="13" style="stroke: #666; fill: none"/><rect x="4.5" y="5.5" width="9" height="8" class="icon-panel-back"/><rect x="4.5" y="5.5" width="9" height="3" class="icon-panel-title"/><rect x="4.5" y="5.5" width="9" height="8" class="icon-border"/>',
+	minimize:
+		'<line x1="3" y1="14.5" x2="15" y2="14.5" class="icon-border" style="stroke-width: 2.5"/>',
+	help: '<path d="M 5,6 A 4,4 0 0 1 10.53,2.3 A 4,4 0 0 1 11.83,8.83 L 9 11.5 L 9 13" class="icon-border" style="stroke-width: 2"/><line x1="9" y1="15" x2="9" y2="17" class="icon-border" style="stroke-width: 2"/>',
+	close: '<line x1="4" y1="4" x2="14" y2="14" class="icon-border" style="stroke-width: 2"/><line x1="4" y1="14" x2="14" y2="4" class="icon-border" style="stroke-width: 2"/>',
+	msgBoxQuestion:
+		'<circle cx="20" cy="20" r="18.5" style="stroke-width: 2; stroke: #04d; fill: #16f"/><path d="M 13,15 A 7,7 0 0 1 22.68,8.53 A 7,7 0 0 1 24.95,19.95 L 20 25 L 20 28" style="stroke-width: 2; stroke: #fff; fill: none"/><line x1="20" y1="31" x2="20" y2="34" style="stroke-width: 2; stroke: #fff; fill: none"/>',
+	msgBoxExclamation:
+		'<polygon points="19,2 21,2 38,36 37,38 3,38 2,36" style="stroke-width: 2; stroke: #a91; fill: #ec2"/><line x1="20" y1="10" x2="20" y2="28" style="stroke-width: 2; stroke: #000; fill: none"/><line x1="20" y1="31" x2="20" y2="34" style="stroke-width: 2; stroke: #000; fill: none"/>',
+	newDocument:
+		'<polygon points="4.5,7.5 8.5,3.5 14.5,3.5 14.5,16.5 4.5,16.5" style="stroke: #333; fill: #fff"/><polyline points="4.5,7.5 8.5,7.5 8.5,3.5" style="stroke: #333; fill: none"/><circle cx="14" cy="14" r="4.75" style="stroke: #030; fill: #0a0"/><line x1="14" y1="17" x2="14" y2="11" style="stroke-width: 2; stroke: #fff; fill: none"/><line x1="11" y1="14" x2="17" y2="14" style="stroke-width: 2; stroke: #fff; fill: none"/>',
+	openDocument:
+		'<polygon points="2.5,4.5 7.5,4.5 9.5,6.5 15.5,6.5 15.5,15.5 2.5,15.5" style="fill: #ec5; stroke: #330"/><polygon points="5.5,8.5 17.5,8.5 15.5,15.5 3.5,15.5" style="fill: #fd6; stroke: #330"/>',
+	saveDocument:
+		'<polygon points="3.5,3.5 16.5,3.5 16.5,16.5 5.5,16.5 3.5,14.5" style="stroke: #139; fill: #36d"/><rect x="7" y="11" width="6" height="5" style="stroke: none; fill: #eef"/><rect x="8" y="12" width="2" height="3" style="stroke: none; fill: #36d"/><rect x="6" y="4" width="8" height="5" style="stroke: none; fill: #fff"/>',
+	saveDocumentAs:
+		'<polygon points="3.5,3.5 16.5,3.5 16.5,16.5 5.5,16.5 3.5,14.5" style="stroke: #139; fill: #36d"/><rect x="7" y="11" width="6" height="5" style="stroke: none; fill: #eef"/><rect x="8" y="12" width="2" height="3" style="stroke: none; fill: #36d"/><rect x="6" y="4" width="8" height="5" style="stroke: none; fill: #fff"/><line x1="8" y1="8" x2="18" y2="18" style="stroke-width: 3; stroke: #0005; fill: none"/><polygon points="8,6 11,7 9,9" style="stroke: none; fill: #fc9"/><polygon points="8,6 9.5,6.5 8.5,7.5" style="stroke: none; fill: #000"/><line x1="10" y1="8" x2="18" y2="16" style="stroke-width: 3; stroke: #000; fill: none"/><line x1="10" y1="8" x2="18" y2="16" style="stroke-width: 2; stroke: #dd0; fill: none"/><polygon points="8,6 9.5,6.5 8.5,7.5" style="stroke: none; fill: #000"/><path d="M 19.06,14.94 A 1.5,1.5 0 0 1 19.06,17.06 A 1.5,1.5 0 0 1 16.94,17.06" style="stroke: none; fill: #000"/><circle cx="18" cy="16" r="1" style="stroke: none; fill: #f33"/>',
+	run: '<polygon points="5,5 15,10 5,15" class="icon-green-fill"/>',
+	interrupt:
+		'<rect x="5" y="5" width="4" height="10" class="icon-red-fill"/><rect x="11" y="5" width="4" height="10" class="icon-red-fill"/>',
+	reset: '<rect x="5" y="5" width="10" height="10" class="icon-red-fill"/>',
+	stepInto:
+		'<rect x="10" y="3" width="7" height="2" class="icon-neutral-fill"/><rect x="13" y="6" width="4" height="2" class="icon-neutral-fill"/><rect x="13" y="9" width="4" height="2" class="icon-neutral-fill"/><rect x="13" y="12" width="4" height="2" class="icon-neutral-fill"/><rect x="10" y="15" width="7" height="2" class="icon-neutral-fill"/><polyline points="8,4 3,4 3,10 6,10" class="icon-blue-line"/><polygon points="5,7 5,13 9.5,10" class="icon-blue-fill"/>',
+	stepOver:
+		'<rect x="10" y="3" width="7" height="2" class="icon-neutral-fill"/><rect x="13" y="6" width="4" height="2" class="icon-neutral-fill"/><rect x="13" y="9" width="4" height="2" class="icon-neutral-fill"/><rect x="13" y="12" width="4" height="2" class="icon-neutral-fill"/><rect x="10" y="15" width="7" height="2" class="icon-neutral-fill"/><polyline points="8,4 3,4 3,16 6,16" class="icon-blue-line"/><polygon points="5,13 5,19 9.5,16" class="icon-blue-fill"/>',
+	stepOut:
+		'<rect x="10" y="3" width="7" height="2" class="icon-neutral-fill"/><rect x="13" y="6" width="4" height="2" class="icon-neutral-fill"/><rect x="13" y="9" width="4" height="2" class="icon-neutral-fill"/><rect x="13" y="12" width="4" height="2" class="icon-neutral-fill"/><rect x="10" y="15" width="7" height="2" class="icon-neutral-fill"/><polyline points="11,10 3,10 3,16 6,16" class="icon-blue-line"/><polygon points="5,13 5,19 9.5,16" class="icon-blue-fill"/>',
+	breakPoint: '<circle cx="10" cy="10" r="3.9" class="icon-red-fill"/>',
+	export: '<polygon points="3,7 10,7 10,3 17,10 10,17 10,13 3,13" class="icon-green-frame"/>',
+	config: '<circle cx="10" cy="10" r="2" class="icon-neutral-fill"/><circle cx="10" cy="10" r="5.7" class="icon-neutral-line" style="stroke-width: 2"/><line x1="15.8" y1="11.55" x2="19.08" y2="12.43" class="icon-neutral-line" style="stroke-width: 2"/><line x1="14.24" y1="14.24" x2="16.65" y2="16.65" class="icon-neutral-line" style="stroke-width: 2"/><line x1="11.55" y1="15.8" x2="12.43" y2="19.08" class="icon-neutral-line" style="stroke-width: 2"/><line x1="8.45" y1="15.8" x2="7.57" y2="19.08" class="icon-neutral-line" style="stroke-width: 2"/><line x1="5.76" y1="14.24" x2="3.35" y2="16.65" class="icon-neutral-line" style="stroke-width: 2"/><line x1="4.2" y1="11.55" x2="0.92" y2="12.43" class="icon-neutral-line" style="stroke-width: 2"/><line x1="4.2" y1="8.45" x2="0.92" y2="7.57" class="icon-neutral-line" style="stroke-width: 2"/><line x1="5.76" y1="5.76" x2="3.35" y2="3.35" class="icon-neutral-line" style="stroke-width: 2"/><line x1="8.45" y1="4.2" x2="7.57" y2="0.92" class="icon-neutral-line" style="stroke-width: 2"/><line x1="11.55" y1="4.2" x2="12.43" y2="0.92" class="icon-neutral-line" style="stroke-width: 2"/><line x1="14.24" y1="5.76" x2="16.65" y2="3.35" class="icon-neutral-line" style="stroke-width: 2"/><line x1="15.8" y1="8.45" x2="19.08" y2="7.57" class="icon-neutral-line" style="stroke-width: 2"/>',
+	restorePanels:
+		'<rect x="2.5" y="2.5" width="15" height="15" style="stroke: #aaa; fill: #fff"/><rect x="11" y="3" width="1" height="14" style="stroke: none; fill: #ccc"/><rect x="3" y="3" width="14" height="1" style="stroke: none; fill: #77f"/><rect x="3" y="12" width="8" height="1" style="stroke: none; fill: #77f"/><path d="M 6.67,7.67 A 4,4 0 0 1 13.39,9.57 A 4,4 0 0 1 8.26,14.3" style="stroke: #222; stroke-width: 1.7; fill: none"/><polygon points="5,5 5,10 10,10" style="stroke: none; fill: #222"/>',
+	editor: '<polygon points="4.5,7.5 8.5,3.5 14.5,3.5 14.5,16.5 4.5,16.5" style="stroke: #333; fill: #fff"/><polyline points="4.5,7.5 8.5,7.5 8.5,3.5" style="stroke: #333; fill: none"/><rect x="10" y="5" width="3" height="1" style="stroke: none; fill: #777"/><rect x="10" y="7" width="2" height="1" style="stroke: none; fill: #777"/><rect x="7" y="9" width="4" height="1" style="stroke: none; fill: #777"/><rect x="6" y="11" width="7" height="1" style="stroke: none; fill: #777"/><rect x="9" y="13" width="4" height="1" style="stroke: none; fill: #777"/>',
+	messages:
+		'<path d="M 4.87,13 A 7,6 0 0 1 12.68,3.15 A 7,6 0 0 1 7.97,14.36 L 4.5 17 Z" style="stroke: #222; fill: #fff"/><rect x="8" y="6" width="3" height="1" style="stroke: none; fill: #777"/><rect x="6" y="8" width="2" height="1" style="stroke: none; fill: #777"/><rect x="9" y="8" width="5" height="1" style="stroke: none; fill: #777"/><rect x="7" y="10" width="4" height="1" style="stroke: none; fill: #777"/>',
+	stackView:
+		'<polygon points="4,5.5 10,8.5 16,5.5 10,2.5" style="stroke: none; fill: #fff"/><polygon points="4,5.5 4,14.5 10,17.5 10,8.5" style="stroke: none; fill: #bbb"/><polygon points="10,17.5 16,14.5 16,5.5 10,8.5" style="stroke: none; fill: #999"/><polyline points="3,8.5 10,11.5 17,8.5" style="stroke-width: 0.6; stroke: #222; fill: none"/><polyline points="3,11.5 10,14.5 17,11.5" style="stroke-width: 0.6; stroke: #222; fill: none"/><polyline points="3,14.5 10,17.5 17,14.5" style="stroke-width: 0.6; stroke: #222; fill: none"/><polygon points="3.5,5.3 3.5,5.7 10,8.5 16.5,5.7 16.5,5.3 10,2.5" style="stroke-width: 0.6; stroke: #222; fill: none"/>',
+	programView:
+		'<polygon points="3,2 13,2 13,5 15,5 15,8 17,8 17,12 13,12 13,14 15,14 15,18 5,18 5,15 3,15 3,11 7,11 7,9 5,9 5,6 3,6" style="stroke:none; fill: #eeeeeec0"/><rect x="4" y="3" width="8" height="2" style="stroke:none; fill: #000"/><rect x="6" y="6" width="8" height="2" style="stroke:none; fill: #000"/><rect x="8" y="9" width="8" height="2" style="stroke:none; fill: #000"/><rect x="4" y="12" width="8" height="2" style="stroke:none; fill: #000"/><rect x="6" y="15" width="8" height="2" style="stroke:none; fill: #000"/>',
+	turtle: '<path d="M 11.5,5 A 2,2.5 0 0 1 7.5,5 A 2,2.5 0 0 1 11.5,5 Z" style="stroke: #070; fill: #2c3"/><polyline points="3.5,6 9.5,11 15.5,6" style="stroke-width: 1.6; stroke: #070; fill: none"/><polyline points="4.5,17 9.5,10 14.5,17" style="stroke-width: 1.6; stroke: #070; fill: none"/><line x1="9.5" y1="17" x2="8.5" y2="19" style="stroke-width: 1.3; stroke: #070; fill: none"/><path d="M 13.5,11.5 A 4,5 0 0 1 5.5,11.5 A 4,5 0 0 1 13.5,11.5 Z" style="stroke: #070; fill: #2c3"/><path d="M 11.15,6.75 A 4,5 0 0 1 12.04,13.74 A 4,5 0 0 1 6.45,14.85" style="stroke: #0709; fill: none"/>',
+	canvas: '<rect x="3" y="2" width="14" height="16" style="stroke: none; fill: #333"/><rect x="4" y="3" width="12" height="14" style="stroke: none; fill: #fff"/><rect x="5.5" y="5.5" width="6" height="6" style="stroke: #00cc; fill: #00c8"/><circle cx="11" cy="11" r="3.5" style="stroke: #c00c; fill: #c008"/>',
+};
